Gerekli düzenlemeler

+++ src/main/java/com/hrms/dataAccess/abstracts/CityDao.java
@@ -3,6 +3,6 @@ package com.hrms.dataAccess.abstracts;
 import com.hrms.entities.concretes.City;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CityDao extends JpaRepository<City, Long> {
+public interface CityDao extends JpaRepository<City, Integer> {
 
 }
