+++ src/main/java/com/hrms/dataAccess/abstracts/JobPositionDao.java
@@ -3,7 +3,7 @@ package com.hrms.dataAccess.abstracts;
 import com.hrms.entities.concretes.JobPosition;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobPositionDao extends JpaRepository<JobPosition, Long> {
+public interface JobPositionDao extends JpaRepository<JobPosition, Integer> {
 
     JobPosition existsByTitle(String title);
