+++ src/main/java/com/hrms/dataAccess/abstracts/WorkingPlaceDao.java
@@ -3,5 +3,5 @@ package com.hrms.dataAccess.abstracts;
 import com.hrms.entities.concretes.WorkingPlace;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkingPlaceDao extends JpaRepository<WorkingPlace, Long> {
+public interface WorkingPlaceDao extends JpaRepository<WorkingPlace, Integer> {
 }
