+++ src/main/java/com/hrms/dataAccess/abstracts/WorkingTimeDao.java
@@ -3,5 +3,5 @@ package com.hrms.dataAccess.abstracts;
 import com.hrms.entities.concretes.WorkingTime;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkingTimeDao extends JpaRepository<WorkingTime, Long> {
+public interface WorkingTimeDao extends JpaRepository<WorkingTime, Integer> {
 }
